refactor)component): move icon component
