+++ src/lib/components/index.ts
@@ -8,7 +8,7 @@ export { Dialog } from "./dialog";
 export { Drawer } from "./drawer";
 export * from "./gallery";
 export { Gallery, ImageFullscreen } from "./gallery";
-export { default as Icon } from "./icons/Icon.svelte";
+export { Icon, SVG } from "./svg";
 export { Image } from "./image";
 export { InputColor } from "./input-color";
 export { InputNumber } from "./input-number";
@@ -25,7 +25,6 @@ export * as Radio from "./radio";
 export { Range } from "./input-range";
 export { ScrollToTop } from "./scroll-to-top";
 export { Sketch, SketchControls } from "./sketch";
-export { default as SVG } from "./SVG.svelte";
 export { Switch, SwitchGroup } from "./switch";
 export { TOC } from "./toc";
 export { default as WorkInProgress } from "./WorkInProgress.svelte";
