+++ src/lib/components/svg/icon.svelte
