+++ src/lib/components/svg/index.ts
@@ -1,0 +1,2 @@
+export { default as Icon } from "./icon.svelte";
+export { default as SVG } from "./svg.svelte";
