+++ src/lib/components/switch/SwitchGroup.svelte
@@ -5,7 +5,7 @@
 <script lang="ts">
 	import { createEventDispatcher } from "svelte";
 	import Switch from "./Switch.svelte";
-	import Icon from "../icons/Icon.svelte";
+	import { Icon } from "../svg";
 	import styles from "./switch-group.module.css";
 
 	export let legend: string;
